+++ opt-env-conf/default.nix
@@ -2,18 +2,18 @@
 , autodocodec-schema, autodocodec-yaml, base, containers, hashable
 , lib, mtl, path, path-io, safe-coloured-text
 , safe-coloured-text-layout, safe-coloured-text-terminfo, selective
-, text, validity, validity-containers
+, text, validity, validity-containers, validity-text
 }:
 mkDerivation {
   pname = "opt-env-conf";
-  version = "0.12.2.0";
+  version = "0.13.0.0";
   src = ./.;
   libraryHaskellDepends = [
     aeson autodocodec autodocodec-nix autodocodec-schema
     autodocodec-yaml base containers hashable mtl path path-io
     safe-coloured-text safe-coloured-text-layout
     safe-coloured-text-terminfo selective text validity
-    validity-containers
+    validity-containers validity-text
   ];
   homepage = "https://github.com/NorfairKing/opt-env-conf#readme";
   description = "Settings parsing for Haskell: command-line arguments, environment variables, and configuration values";